Add files via upload
modified files to make main.py smaller
