+++ main.py
@@ -1,32 +1,10 @@
-from googleapiclient.discovery import build
-
-import os
-
 from dict_to_list import get_all_values
 
-# paste the API key generated by you here
-
-# The API Key here is hidden by saving to an environment variable.
-API_KEY = os.environ.get('API_KEY')
+from youtube_build import request
 
 # Sana, Fauna, Kronii, Mumei, Baelz Channel URLs
 channel_url_id = ["UCsUj0dszADCGbF3gNrQEuSQ", "UCO_aKKYxn4tvrqPjcTzZ6EQ",
               "UCmbs8T6MWqUHP1tIQvSgKrg", "UC3n5uGu18FoCy23ggWWp8tA", "UCgmPnx-EEeOrZSg5Tiw7ZRQ"]
-
-# for building a request to the YouTube API using an API Key.
-youtube = build('youtube', 'v3', developerKey=API_KEY)
-
-
-# Request to be sent to the Youtube API. The Channel ID is obtained from the channel_url_id list defined earlier.
-# the request is made as per the Youtube API documentation.
-def request(channel_id):
-    request_to_channel = youtube.channels().list(
-        part='statistics',
-        id=channel_id
-    )
-    response_to_request = request_to_channel.execute()
-    return response_to_request
-
 
 # printing out the response output obtained for the request.
 for channel in channel_url_id:
